+++ snippets/usage.rs
@@ -6,7 +6,8 @@ compose_idents!(
     my_const = [upper(foo), _, lower(BAR)];
     my_static = [upper(lower(BAR))];
     MY_SNAKE_CASE_STATIC = [snake_case(snakeCase)];
-    MY_CAMEL_CASE_STATIC = [camel_case(camel_case)]; {
+    MY_CAMEL_CASE_STATIC = [camel_case(camel_case)];
+    MY_UNIQUE_STATIC = [hash(0b11001010010111)]; {
     fn my_fn_1() -> u32 {
         123
     }
@@ -19,6 +20,7 @@ compose_idents!(
     static my_static: u32 = 42;
     static MY_SNAKE_CASE_STATIC: u32 = 42;
     static MY_CAMEL_CASE_STATIC: u32 = 42;
+    static MY_UNIQUE_STATIC: u32 = 42;
 });
 
 macro_rules! outer_macro {
@@ -32,6 +34,20 @@ macro_rules! outer_macro {
 }
 
 outer_macro!(foo);
+
+macro_rules! global_var_macro {
+    () => {
+        // `my_static` is going to be unique in each invocation of `global_var_macro!()`.
+        // But within the same invocation `hash(1)` will yield the same result.
+        compose_idents!(
+            my_static = [foo, _, hash(1)]; {
+            static my_static: u32 = 42;
+        });
+    };
+}
+
+global_var_macro!();
+global_var_macro!();
 
 assert_eq!(foo_baz(), 123);
 assert_eq!(spam_1_eggs(), 321);