+++ src/funcs.rs
@@ -1,3 +1,6 @@
+use crate::core::State;
+use std::hash::{DefaultHasher, Hash, Hasher};
+
 /// Converts the input string to snake_case.
 pub fn to_snake_case(input: &str) -> String {
     let mut result = String::new();
@@ -38,10 +41,24 @@ pub fn to_camel_case(input: &str) -> String {
     result
 }
 
+/// Generates an identifier from a provided seed deterministically within a single macro invocation.
+///
+/// `hash(1)` called within a single macro invocation will always return the same
+/// value but different in another macro invocation.
+pub fn hash(input: &str, state: &State) -> String {
+    let mut hasher = DefaultHasher::new();
+    state.seed.hash(&mut hasher);
+    input.hash(&mut hasher);
+    let hash = hasher.finish().to_string();
+    let result = format!("__{}", hash);
+    result
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
     use rstest::rstest;
+    use syn::Ident;
 
     #[rstest]
     #[case("fooBar", "foo_bar")]
@@ -71,5 +88,29 @@ mod tests {
     fn test_to_camel_case(#[case] input: &str, #[case] expected: &str) {
         let actual = to_camel_case(input);
         assert_eq!(actual, expected, "Input: {}", input);
+    }
+
+    #[rstest]
+    fn test_random_valid_ident() {
+        let state = State { seed: 1 };
+        let actual = hash("1", &state);
+        let ident_result = syn::parse_str::<Ident>(actual.as_str());
+
+        assert!(
+            ident_result.is_ok(),
+            "Result: {},\nError: {}",
+            actual,
+            ident_result.unwrap_err(),
+        );
+    }
+
+    #[rstest]
+    fn test_random_determinism() {
+        let state = State { seed: 1 };
+        let expected = hash("1", &state);
+        let actual = hash("1", &state);
+
+        assert_eq!(actual, expected);
+        assert_ne!(hash("2", &state), expected);
     }
 }