+++ src/lib.rs
@@ -4,7 +4,7 @@ mod core;
 mod eval;
 mod funcs;
 
-use crate::core::Expr;
+use crate::core::{Expr, State};
 use crate::eval::Eval;
 use proc_macro::TokenStream;
 use quote::{format_ident, quote};
@@ -23,9 +23,9 @@ struct IdentSpecItem {
 }
 
 impl IdentSpecItem {
-    fn replacement(&self) -> Ident {
+    fn replacement(&self, state: &State) -> Ident {
         let ident = self.exprs.iter().fold("".to_string(), |acc, item| {
-            format!("{}{}", acc, item.eval())
+            format!("{}{}", acc, item.eval(state))
         });
         format_ident!("{}", ident)
     }
@@ -65,10 +65,10 @@ impl Parse for IdentSpec {
 }
 
 impl IdentSpec {
-    fn replacements(&self) -> HashMap<Ident, Ident> {
+    fn replacements(&self, state: &State) -> HashMap<Ident, Ident> {
         self.items
             .iter()
-            .map(|item| (item.alias.clone(), item.replacement()))
+            .map(|item| (item.alias.clone(), item.replacement(state)))
             .collect()
     }
 }
@@ -98,6 +98,8 @@ impl VisitMut for ComposeIdentsVisitor {
     }
 }
 
+static mut COUNTER: u64 = 0;
+
 /// Compose identifiers from the provided parts and replace their aliases in the code block.
 ///
 /// # Example
@@ -107,9 +109,15 @@ impl VisitMut for ComposeIdentsVisitor {
 /// ```
 #[proc_macro]
 pub fn compose_idents(input: TokenStream) -> TokenStream {
+    let state = State {
+        seed: unsafe {
+            COUNTER += 1;
+            COUNTER
+        },
+    };
     let args = parse_macro_input!(input as ComposeIdentsArgs);
     let mut visitor = ComposeIdentsVisitor {
-        replacements: args.spec.replacements(),
+        replacements: args.spec.replacements(&state),
     };
     let mut block = args.block;
     visitor.visit_block_mut(&mut block);