+++ tests/tests.rs
@@ -14,4 +14,5 @@ fn compile_tests() {
     t.pass("tests/compile/funcs/different_arg_types.rs");
     t.pass("tests/compile/funcs/snake_case.rs");
     t.pass("tests/compile/funcs/camel_case.rs");
+    t.pass("tests/compile/funcs/hash.rs");
 }